Wait for target interface state when adding/removing NM connections [VPNLINUX-748]
Previously we considered the KS was enabled as soon as the underlying NM
connection was added, and disabled as soon as it was removed.

However, the dummy interface that NM creates/removes when enabling/disabling
the KS was not ready instantly after the NM connection was added/removed. This
was leading to issues that were previously workarounded by adding a delay
before starting the VPN connection.

This commit fixes this situation by monitoring the state of the dummy interface
before acknowledging that the KS has been enabled/disabled. The KS is only
considered enabled when the dummy interface reaches the "activated" state.
Similarly, the KS is only considered disabled when the dummy interface reaches
the "disconnected" state. This applies to all KS types (full, routed and IPv6).